moved symbols
